# back/docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /quests/{id}:
    get:
      operationId: QuestsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: Quest by id
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 1
                  title: Rescue the Merchant
                  description: Escort the merchant safely to the city.
                  statusId: 1
                  createdAt: '2025-10-30T12:00:00.000Z'
                  updatedAt: '2025-10-30T12:34:56.000Z'
      security:
        - bearer: []
      tags:
        - Quests
    patch:
      operationId: QuestsController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 42
            type: number
      requestBody:
        required: true
        description: Fields to update (partial)
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                name: Rescue the Merchant (Hard)
                description: Hard mode variant.
                reward: 750
      responses:
        '200':
          description: Updated quest
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 42
                  name: Rescue the Merchant (Hard)
                  description: Hard mode variant.
                  reward: 750
                  statusId: 1
                  createdAt: '2025-10-30T12:00:00.000Z'
                  updatedAt: '2025-10-30T12:45:00.000Z'
      security:
        - bearer: []
      tags:
        - Quests
...
  /equipment/{id}:
    get:
      operationId: EquipmentController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Equipment ID
          schema:
            example: 10
            type: number
      responses:
        '200':
          description: Equipment by id
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 10
                  name:
                    type: string
                    example: Longsword
                  equipmentTypeId:
                    type: number
                    example: 7
                  cost:
                    type: number
                    example: 100
                  maxDurability:
                    type: number
                    example: 50
                  currentDurability:
                    type: number
                    example: 50
      security:
        - bearer: []
      tags:
        - Equipment
    patch:
      operationId: EquipmentController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Equipment ID
          schema:
            example: 10
            type: number
      requestBody:
        required: true
        description: Fields to update (partial)
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Enchanted Longsword
                equipmentTypeId:
                  type: number
                  example: 8
                cost:
                  type: number
                  example: 150
                maxDurability:
                  type: number
                  example: 60
              additionalProperties: false
      responses:
        '200':
          description: Updated equipment
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 10
                  name:
                    type: string
                    example: Enchanted Longsword
                  equipmentTypeId:
                    type: number
                    example: 8
                  cost:
                    type: number
                    example: 150
                  maxDurability:
                    type: number
                    example: 60
                  currentDurability:
                    type: number
                    example: 45
      security:
        - bearer: []
      tags:
        - Equipment
    delete:
      operationId: EquipmentController_delete
      parameters:
        - name: id
          required: true
          in: path
          description: Equipment ID
          schema:
            example: 10
            type: number
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 10
                  deleted:
                    type: boolean
                    example: true
      security:
        - bearer: []
      tags:
        - Equipment
...
  /quest-stock-equipment:
    get:
      operationId: QuestStockEquipmentController_list
      parameters:
        - name: questId
          required: false
          in: query
          description: Filter by quest ID
          schema:
            type: number
            example: 12
      responses:
        '200':
          description: List of quest-stock-equipment
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                      example: 101
                    questId:
                      type: number
                      example: 12
                    equipmentStockId:
                      type: number
                      example: 3
                    createdAt:
                      type: string
                      format: date-time
                      example: '2025-10-30T12:00:00.000Z'
                    updatedAt:
                      type: string
                      format: date-time
                      example: '2025-10-30T12:34:56.000Z'
      security:
        - bearer: []
      tags:
        - Quest Stock Equipment
    post:
      operationId: QuestStockEquipmentController_attach
      parameters: []
      requestBody:
        required: true
        description: Attach equipment stock to a quest
        content:
          application/json:
            schema:
              type: object
              properties:
                questId:
                  type: number
                  example: 12
                equipmentStockId:
                  type: number
                  example: 3
              required:
                - questId
                - equipmentStockId
              additionalProperties: false
      responses:
        '201':
          description: Attachment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 101
                  questId:
                    type: number
                    example: 12
                  equipmentStockId:
                    type: number
                    example: 3
                  createdAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:00:00.000Z'
                  updatedAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:00:00.000Z'
      security:
        - bearer: []
      tags:
        - Quest Stock Equipment
  /quest-stock-equipment/{id}:
    delete:
      operationId: QuestStockEquipmentController_delete
      parameters:
        - name: id
          required: true
          in: path
          description: Attachment ID
          schema:
            example: 101
            type: number
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 101
                  deleted:
                    type: boolean
                    example: true
      security:
        - bearer: []
      tags:
        - Quest Stock Equipment
...
info:
  title: API "Les kips AH"
  description: |-
    This is the API information regarding the project "Les capuches d'Opale".

    C'est un outil de gestion d'aventurier pour une guilde.
  version: 1.0.0
  contact: {}
  termsOfService: https://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http